requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2
asyncpg
pydantic
pandas
gspread
oauth2client
python-dotenv
aiohttp
